Move AddressAutocompleteInput to form subdirectory and update imports
